requirements.txt
anthropic==0.89.0
flask==3.1.3
python-dotenv==1.2.1
gunicorn==23.0.0
